# packages/webui/frontend/src/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "linling-webui",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "description": "Return events from one or all visible bots' ring buffers.",
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bot Id"
            }
          },
          {
            "name": "since_seq",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Seq"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "scope_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Kind"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{bot_id}/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "operationId": "get_event_api_events__bot_id___event_id__get",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{bot_id}/{event_id}/replay": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Replay Event",
        "description": "Re-dispatch a buffered event through the router in dry-run mode.\n\nOnly dry-run is supported (``dry_run=True``). Actual re-sending to an\nadapter would bypass rate-limits and idempotency — we refuse.",
        "operationId": "replay_event_api_events__bot_id___event_id__replay_post",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventEnvelope": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "bot_id": {
            "type": "string",
            "title": "Bot Id"
          },
          "scope": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scope"
          },
          "sender": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sender"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "segments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Segments"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "id",
          "platform",
          "bot_id",
          "scope",
          "sender",
          "time",
          "kind",
          "segments",
          "text"
        ],
        "title": "EventEnvelope",
        "description": "Slimmed event view sent to the browser.\n\nAvoids returning the raw adapter payload (may contain secrets). The\nfull pydantic Event is not exposed; adapters own `.raw`."
      },
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EventPage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_ReplayRequest": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "_ReplayRequest"
      },
      "_ReplayResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "title": "_ReplayResponse"
      }
    }
  }
}